skynet_linker_cli/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# Redis
redis==5.0.1

# Cryptography
PyNaCl==1.5.0

# CLI
typer[all]==0.9.0
rich==13.7.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# JSON schema validation
jsonschema==4.20.0

# HTTP client
httpx==0.26.0

# YAML support
PyYAML==6.0.1

# Async support
asyncio==3.4.3
aioredis==2.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Logging
python-json-logger==2.0.7
